+++ internal/config/add-account.go
@@ -20,11 +20,11 @@ package config
 
 import (
 	"fmt"
-	output2 "github.com/onflow/flow-cli/internal/output"
 
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/internal/command"
+	"github.com/onflow/flow-cli/internal/prompt"
 	"github.com/onflow/flow-cli/pkg/flowkit"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
 	"github.com/onflow/flow-cli/pkg/flowkit/output"
@@ -65,7 +65,7 @@ func addAccount(
 	}
 
 	if !flagsProvided {
-		accountData = output2.NewAccountPrompt()
+		accountData = prompt.NewAccountPrompt()
 	}
 
 	account, err := config.StringToAccount(
