+++ internal/config/add-contract.go
@@ -20,11 +20,10 @@ package config
 
 import (
 	"fmt"
-	output2 "github.com/onflow/flow-cli/internal/output"
-
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/internal/command"
+	"github.com/onflow/flow-cli/internal/prompt"
 	"github.com/onflow/flow-cli/pkg/flowkit"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
 	"github.com/onflow/flow-cli/pkg/flowkit/output"
@@ -64,7 +63,7 @@ func addContract(
 	}
 
 	if !flagsProvided {
-		contractData = output2.NewContractPrompt()
+		contractData = prompt.NewContractPrompt()
 	}
 
 	contract := config.StringToContract(
