+++ internal/config/add-deployment.go
@@ -20,11 +20,10 @@ package config
 
 import (
 	"fmt"
-	output2 "github.com/onflow/flow-cli/internal/output"
-
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/internal/command"
+	"github.com/onflow/flow-cli/internal/prompt"
 	"github.com/onflow/flow-cli/pkg/flowkit"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
 	"github.com/onflow/flow-cli/pkg/flowkit/output"
@@ -62,7 +61,7 @@ func addDeployment(
 	}
 
 	if !flagsProvided {
-		deployData = output2.NewDeploymentPrompt(*state.Networks(), state.Config().Accounts, *state.Contracts())
+		deployData = prompt.NewDeploymentPrompt(*state.Networks(), state.Config().Accounts, *state.Contracts())
 	}
 
 	deployment := config.StringToDeployment(
