+++ internal/config/remove-account.go
@@ -19,7 +19,7 @@
 package config
 
 import (
-	output2 "github.com/onflow/flow-cli/internal/output"
+	"github.com/onflow/flow-cli/internal/prompt"
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/internal/command"
@@ -53,7 +53,7 @@ func removeAccount(
 	if len(args) == 1 {
 		name = args[0]
 	} else {
-		name = output2.RemoveAccountPrompt(state.Config().Accounts)
+		name = prompt.RemoveAccountPrompt(state.Config().Accounts)
 	}
 
 	err := state.Accounts().Remove(name)
