+++ internal/super/setup.go
@@ -22,7 +22,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	output2 "github.com/onflow/flow-cli/internal/output"
+	"github.com/onflow/flow-cli/internal/prompt"
 	"io"
 	"net/http"
 	"os"
@@ -94,7 +94,7 @@ func create(
 			scaffoldList[i] = fmt.Sprintf("%s - %s", output.Bold(s.Name), s.Description)
 		}
 
-		selected := output2.ScaffoldPrompt(scaffoldList)
+		selected := prompt.ScaffoldPrompt(scaffoldList)
 		pickedScaffold = scaffolds[selected]
 	}
